Writing word in lower case now

+++ src/main/java/com/invertedindex/positivityrank/PositivityRankMapper.java
@@ -11,10 +11,10 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
-public class PositivityRankMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+public class PositivityRankMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
 
     @Override
-    protected void map(LongWritable key, Text value, Mapper<LongWritable, Text, Text, IntWritable>.Context context) throws IOException, InterruptedException {
+    protected void map(LongWritable key, Text value, Mapper<LongWritable, Text, Text, LongWritable>.Context context) throws IOException, InterruptedException {
         Configuration conf = context.getConfiguration();
         List<String> positiveWords = Arrays.asList(conf.getStrings("positiveWords"));
 
@@ -28,7 +28,7 @@ public class PositivityRankMapper extends Mapper<LongWritable, Text, Text, IntWr
                 System.out.println("Word " + word);
                 System.out.println("File and line number " + wordMetadata);
                 String fileName = wordMetadata.split(":=:")[0];
-                context.write(new Text(fileName), new IntWritable(1));
+                context.write(new Text(fileName), new LongWritable(1));
             }
         }
     }
