+++ src/main/java/com/invertedindex/positivityrank/PositivityRankRunner.java
@@ -7,6 +7,7 @@ import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -60,7 +61,7 @@ public class PositivityRankRunner extends Configured implements Tool {
         job.setJarByClass(PositivityRankRunner.class);
 
         job.setMapOutputKeyClass(Text.class);
-        job.setMapOutputValueClass(IntWritable.class);
+        job.setMapOutputValueClass(LongWritable.class);
         job.setMapperClass(PositivityRankMapper.class);
         job.setReducerClass(PositivityRankReducer.class);
